# Conflicts:
#   src/main/java/com/trade_accounting/controllers/rest/EmployeeRestController.java
#   src/main/java/com/trade_accounting/services/impl/EmployeeServiceImpl.java
#   src/main/java/com/trade_accounting/services/impl/fias/StreetServiceImpl.java
#   src/main/java/com/trade_accounting/services/interfaces/EmployeeService.java
#   src/test/java/com/trade_accounting/services/impl/EmployeeServiceImplTest.java

+++ src/main/java/com/trade_accounting/services/impl/fias/StreetServiceImpl.java
@@ -30,7 +30,7 @@ public class StreetServiceImpl implements StreetService {
     }
 
     @Override
-    public StreetDto getById(Integer id) {
+    public StreetDto getById(Long id) {
         Street street = streetRepository.findById(id).orElse(new Street());
         return dtoMapper.toStreetDto(street);
     }
